% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conformal.R
\name{print.somalign_conformal}
\alias{print.somalign_conformal}
\title{Print a somalign_conformal object}
\usage{
\method{print}{somalign_conformal}(x, ...)
}
\arguments{
\item{x}{A \code{somalign_conformal} object.}

\item{...}{Ignored.}
}
\value{
\code{x}, invisibly.
}
\description{
Print a somalign_conformal object
}
