% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.R
\name{print.somalign_exclusion_test}
\alias{print.somalign_exclusion_test}
\title{Print a somalign_exclusion_test object}
\usage{
\method{print}{somalign_exclusion_test}(x, ...)
}
\arguments{
\item{x}{A \code{somalign_exclusion_test} object.}

\item{...}{Ignored.}
}
\value{
\code{x}, invisibly.
}
\description{
Print a somalign_exclusion_test object
}
