% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gromov.R
\name{print.somalign_gw_fit}
\alias{print.somalign_gw_fit}
\title{Print a somalign_gw_fit object}
\usage{
\method{print}{somalign_gw_fit}(x, ...)
}
\arguments{
\item{x}{A \code{somalign_gw_fit} object.}

\item{...}{Ignored.}
}
\value{
\code{x}, invisibly.
}
\description{
Print a somalign_gw_fit object
}
