% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.R
\name{print.somalign_novelty_candidates}
\alias{print.somalign_novelty_candidates}
\title{Print a somalign_novelty_candidates object}
\usage{
\method{print}{somalign_novelty_candidates}(x, ...)
}
\arguments{
\item{x}{A \code{somalign_novelty_candidates} object.}

\item{...}{Ignored.}
}
\value{
\code{x}, invisibly.
}
\description{
Print a somalign_novelty_candidates object
}
