% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.R
\name{print.somalign_query}
\alias{print.somalign_query}
\title{Print a somalign_query object}
\usage{
\method{print}{somalign_query}(x, ...)
}
\arguments{
\item{x}{A \code{somalign_query} object.}

\item{...}{Ignored.}
}
\value{
\code{x}, invisibly.
}
\description{
Print a somalign_query object
}
\examples{
if (requireNamespace("kohonen", quietly = TRUE)) {
  set.seed(1)
  mat <- matrix(rnorm(20), nrow = 10, ncol = 2,
                dimnames = list(NULL, c("F1", "F2")))
  ref <- somalign_train_reference(mat, grid = kohonen::somgrid(2, 2, "hexagonal"),
                                  rlen = 5)
  qry <- somalign_query(mat, ref, grid = kohonen::somgrid(2, 2, "hexagonal"),
                        rlen = 5)
  print(qry)
}
}
