% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sensitivity.R
\name{print.somalign_subspace_sensitivity}
\alias{print.somalign_subspace_sensitivity}
\title{Print a somalign_subspace_sensitivity object}
\usage{
\method{print}{somalign_subspace_sensitivity}(x, ...)
}
\arguments{
\item{x}{A \code{somalign_subspace_sensitivity} object.}

\item{...}{Ignored.}
}
\value{
\code{x}, invisibly.
}
\description{
Print a somalign_subspace_sensitivity object
}
