% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/topology.R
\name{print.somalign_topology}
\alias{print.somalign_topology}
\title{Print a somalign_topology object}
\usage{
\method{print}{somalign_topology}(x, ...)
}
\arguments{
\item{x}{A \code{somalign_topology} object.}

\item{...}{Ignored.}
}
\value{
\code{x}, invisibly.
}
\description{
Print a somalign_topology object
}
