% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.R
\name{print.somalign_fit}
\alias{print.somalign_fit}
\title{Print a somalign_fit object}
\usage{
\method{print}{somalign_fit}(x, ...)
}
\arguments{
\item{x}{A \code{somalign_fit} object.}

\item{...}{Ignored.}
}
\value{
\code{x}, invisibly.
}
\description{
Print a somalign_fit object
}
