% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.R
\name{print.somalign_query}
\alias{print.somalign_query}
\title{Print a somalign_query object}
\usage{
\method{print}{somalign_query}(x, ...)
}
\arguments{
\item{x}{A \code{somalign_query} object.}

\item{...}{Ignored.}
}
\value{
\code{x}, invisibly.
}
\description{
Print a somalign_query object
}
