% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.R
\name{print.somalign_reference}
\alias{print.somalign_reference}
\title{Print a somalign_reference object}
\usage{
\method{print}{somalign_reference}(x, ...)
}
\arguments{
\item{x}{A \code{somalign_reference} object.}

\item{...}{Ignored.}
}
\value{
\code{x}, invisibly.
}
\description{
Print a somalign_reference object
}
