% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/somalign-package.R
\docType{package}
\name{somalign-package}
\alias{somalign}
\alias{somalign-package}
\title{somalign: align query SOMs to fixed reference SOMs}
\description{
The package keeps direct projection into a fixed reference \code{kohonen} SOM as
the conservative primary result and returns auxiliary optimal-transport
corrected projections for visualisation and annotation.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://mdmanurung.github.io/somalign}
  \item \url{https://github.com/mdmanurung/somalign}
  \item Report bugs at \url{https://github.com/mdmanurung/somalign/issues}
}

}
\author{
\strong{Maintainer}: Mikhael Manurung \email{mikhael.manurung@gmail.com}

Authors:
\itemize{
  \item Mikhael Manurung \email{mikhael.manurung@gmail.com}
}

}
\keyword{internal}
